+++ src/oihana/commands/enums/outputs/Palette.php
@@ -2,6 +2,8 @@
 
 namespace oihana\commands\enums\outputs;
 
+use InvalidArgumentException;
+
 use oihana\reflect\traits\ConstantsTrait;
 
 /**
@@ -59,4 +61,30 @@ class Palette
      * @var string White color
      */
     public const string WHITE = 'white';
+
+    /**
+     * Indicates if the passed value is a valid color.
+     * @param mixed $value
+     * @return void
+     */
+    public static function assertColor( mixed $value ) :void
+    {
+        $check = match( $value )
+        {
+            self::BLACK ,
+            self::BLUE,
+            self::CYAN,
+            self::GREEN,
+            self::MAGENTA,
+            self::RED,
+            self::YELLOW,
+            self::WHITE => true  ,
+            default     => false ,
+        };
+
+        if( !$check )
+        {
+            throw new InvalidArgumentException( sprintf( "Invalid foreground color '%s'" , $value ) ) ;
+        }
+    }
 }